(PUP-3186) Skip creation of symlinked default directory env
To provide a better out of the box user experience, Puppet will ensure
the existence of the default production directory environment.  However
if this path has already been created as a symlink, the File resource
will ensure it is replaced with an empty directory.

Since an existing symlink is an indication that the user has already
made configuration changes, and the purpose of ensuring the existence of
the production directory environment is purely for assisting in new,
unmodified installations, this commit skips ensuring anything about the
production directory environment when a symlink is detected.

With Josh Partlow <jpartlow@puppetlabs.com>