(PUP-3162) Switch to a whitelist for $environment interpolation
Previous commit blacklisted a small set of settings from having
$environment interpolated.  After some discussion, we decided to broaden
this to a whitelist, and the only setting we can currently think of
needing $environment interpolation when using directory environments is
config_version.

We also decided to not halt startup for default_manifest with
$environment in it, as these errors are difficult to read with the
current error reporting from a rack master.  So the validation for
default_manifest is removed.