(maint) Generate unique account names
Previously, the windows security_spec could fail if the same test was
executed in parallel, since the test assumed the account didn't exist.

This commit makes the names unique. The maximum length for local
user and groups accounts in 2003 is 20 and 256 characters[1],
respectively, and we remain below that limit in both cases.

[1] http://technet.microsoft.com/en-us/library/cc783323.aspx