(PUP-3136) Fix upstart acceptance test on lucid
Due to the fact that this acceptance test depends on the output
of `service rabbitmq-server status`, it was failing on lucid.
This was due to the fact that lucid has an older version of rabbitmq
and therefor a different output than trusty and precise.

Update the regexes and exit codes so that they support lucid. It's
worth noting that since these tests are based on the text output
of `service rabbitmq-server status`, they are fragile and may break
if rabbitmq changes the output of its init.d script. Additionally, if
rabbitmq switches from init.d to upstart we will need to find another
package.

Paired with Will Hopper <whopper@puppetlabs.com>