(PUP-2558) Remove instrumentation system
Commit d4c80943 added the instrumentation system in December of 2011,
but was never put into use. We've seen transitory issues with the
instrumentation system, autoloading, and the definition of constants,
and in the mean time we've implemented a Profiler system that has been
put to real world use. It does us no good to have this code lying
dormant, so this commit removes it.