(PUP-3245) Profile function calls
The new evaluator implementation has a different code path for function
calls, but this path didn't call out to the profiler to track that the
call was made, this meant that new function calls were not showing up
anywhere.

This updates the call_function runtime to also profile that the call was
made.