(maint) Ensure config acceptance resets uid
Now that we are testing puppet packages with service scripts, the puppet
master defaults to running, and we cannot adjust the puppet user uid
while that user is associated with a runing process. This acceptance
test was stopping the puppet master to remove the system puppet
user/group, but was failing to stop the puppet master when restoring the
system user group, resulting in the puppet system user retaining it's
new uid, and some files (particularly in ssldir) being owned by the old
uid.  This would potentially cause failures depending on test order.

This commit fixes the puppet_manages_own_configuration_in_robust_manner
test to stop the puppet master and restore the original user uid.