(PUP-2560) Convert acceptance test for #7117 to unit test
The acceptance test for Redmine issur #7117 assumed that we would always
be able to reach the inventory the service and blindly did an HTTP
request against that to verify that directives in auth.conf could be
restricted to an environment. This is a very heavyweight way to test
this functionality, and it turns out that removing the inventory service
broke this test anyways. This commit adds a unit test to verify the
behavior of network rights with environments and removes the acceptance
test in favor of the unit test.