(maint) use grep for ssh_authorized_key test
In the CI runs of acceptance/tests/resource/ssh_authorized_key/destroy.rb,
the length of the ~/.ssh/authorized_keys file can be large enough
to cause the step to not complete correctly.

This commit changes the validation to a grep command verifying that
the exit code is 1 indicating that the name no longer is contained
in the file.