(maint) Actually test expected rights behavior
A number of tests were spuriously passing because we were doing a blind
check for truthiness, and `expect(:dunno).to be_true` will pass. In
addition we were checking a negative test that could never fail because
we were passing unmatchable arguments. This commit corrects the tests to
verify the correct behavior.