(maint) Remove the resource/host/should_query test
This test was attempting to verify that the host provider will show an
entry as present in a target host file.  But in fact, because Puppet has
just the single resource representing a host entry of the given name,
what this test is doing is ensuring that the 'localhost' host entry is
found in the given target file. On vms with two localhost entries (say
127.0.0.1 and ::1) this ends up moving the ::1 entry to the target file,
and the system continues to work. But on a system with a single
localhost entry in its /etc/hosts, this ends up moving that to the
target tmp file and crippling localhost name resolution for subsequent
tests. See PUP-1479 for more info.