PUP-3336 split $libdir when updating $LOAD_PATH
As reported, $libdir can (but typically won't) take a multi-part
libdir.  This was being used wholesale in updating $LOAD_PATH, but
ruby have any special treatment, so would just have an invalid path in
$LOAD_PATH.

With this commit we split libdir when updating $LOAD_PATH, to ensure
we're updating each component correctly.