(PUP-3341) force the use of the overridden environment for puppet apply
puppet apply overrides the current environment to use the manifest that was
specified on the command line. An ENC would break out of this environment
by specifying an environment name. The node object that is returned from
the node terminus will override the environment that had been initialised
through the usual means.

We now forcibly set the altered environment after the node has been retrieved.
This makes more sense for apply - the command line should take full control.