(PUP-3277) Change hiera lookup tests to use a temp production direnv
The hiera tests were placing some test data in the default configuration
paths and some in temp directories, and were relying on the manifest and
modulepath settings.  These settings are deprecated and will be removed.
The tests now setup all configuration in a temporary directory for
consistency and isolation, and make use of environmentpath instead.