(PUP-3081) Remove dependencies on specific notice
This family of tests was applying manifests and predicating success
on a specific notice. The wording of the notice was changed which
broke these tests. But this also exposed that testing for this message
wasn't needed anyway - there are other steps in these tests that
do a real check for success/failure.

While I was in there I added some blank lines between test steps,
so that it's not quite such a wall of text. It's a bit more readable
now (I hope).