(PUP-3081) Check for absence of warnings rather than specific notices
For this ticket we changed the Notice associated with catalog application.
This test had a very specific regex it tested against, but the real issue
it was testing for was the absence of warnings rather than precise fidelity
to a particular pair of notices.

So this commit updates the regex to just check for the absence of warnings.