(PUP-3277) Disable PE MCollective group in ssl tests
When the tests are run in PE, the PE MCollective group adds a class for
manging MCollective resources an any PE agent node.  This interferes
with ssl tests which are testing initial cert generation and signing,
because PE attempts to reset the MCollective cert for the agent, and the
MCollective configuration on the agent to match.  By instead using the
classifier_utils helper to temporarily disable this group for the test
nodes, we can test the cert generation in isolation.