(PUP-3277) Fix two acceptance tests
The changes brought in for PUP-3277 were tested with PE, but running
them in our CI brought up several issues, one of which due to the Beaker
create_tmpdir_for_user method not working on windows, and the other
surfacing in use_enc_environment_for_files because directory
environments do not pick up module file changes and the test was
assuming it could change the module manifest on the master and have it
picked up by different agent runs.

Switched to manually chowning the tmpdir with a puppet apply for the
first problem, and setting environment_timeout to 0 for the latter.