(PUP-3352) Remove `zlib` setting
Previously, puppet had a `zlib` setting to control whether to use zlib
compression when connecting to the forge or when using `b64_zlib_yaml`
as a network serialization format.

This commit removes the `zlib` setting, and instead, if ruby supports
the `zlib` library, then we will accept compressed content, e.g. when
downloading modules from the forge.

The b64_zlib_yaml format is being removed in Puppet 4, so that part
doesn't matter.

The `zlib` setting defaulted to false, so there shouldn't be any
behavior changes other than you can no longer set the property to
false.