(PUP-3352) Always accept compressed content from the master
Previously, a puppet agent would not use compression when downloading
catalogs and file contents. Compression could be enabled via the
`http_compression` setting, though it defaulted to false.

This commit removes the `http_compression` setting. If the agent ruby
supports zlib, then the agent will add an Accept-Encoding header
specifying that it will accept compress content.

If the agent ruby doesn't support zlib, then the agent will not set
the header, and it will behave like it used to.

Note that not all HTTP servers support compression, e.g. webrick, or are
configured to serve compressed content, e.g. passenger. So in order for
the agent to receive compressed content, the agent must accept it, and
the server must be configured to send it.