(PUP-3277) Filter group properties before PUT
Originally the helper would roundtrip the MCollective group retrieved
via GET, merely updating the 'rule' property. We had a failure where the
MCollective group came back with a 'deleted' object in its properties,
which is not a property you can PUT to a group. I've changed the helper
to select a set of allowed properties that we can PUT back to the
original group.