(PUP-3277) Use classifier enc for some enc tests on pe
The store_configs/enc_provides_node_when_storeconfigs_enabled.rb test
checks that storeconfigs does not prevent an enc from working. Since
FOSS runs default to no enc or storeconfigs, the test would setup both;
but when running in PE, both are already available and the test needed
to be updated to accomodate that.

In addition changes were made to use the production directory
environment. The tests/environment/use_enc_environment.rb test was also
updated to make use of the classifier when testing against PE.