(PUP-643) Ignore pkg's certificate expiration warning messages
When a pkg certificate is in it's last 30 days before expiration, the
pkg tools emit a warning message so you can get a new certificate for
some operations. The 'latest' method issues a 'pkg -Hn <package>' which
trips this warning message, but the message was unexpected and 'latest'
would raise, failing the catalog application.

This commit simply ignores such certificate expiration warnings.

While I was in there, I split a very compact line of code across a couple
lines and added some spaces, in the name of readability.