(maint) Remove the agent role from the master node on Solaris 11
Background: the Solaris 11 acceptance tests take a long time. One reason
is that the agent role is executed on both the agent and master node,
doubling the time taken for agent testing. This isn't giving us more
coverage and it increases the test feedback loop.

So this commit drops the agent role from the master node.