(PUP-2923) Add support for assumeyes to yum provider
Previously when yum needs to prompt the user for confirmation
it would fail because the assumeyes flag was not set.

Example: When trying to set gpgkey attribute to url
or having the repo_gpgcheck attribute set to 1 would cause yum
to prompt the user for confirmation on importing a GPG key. This
would cause puppet to fail.

This provides the ability to set the assumeyes attribute, defaulting
to not set.