(PUP-3449) Skip merging node parameters reserved in top scope
The top scope has 'trusted' and 'facts' as reserved variable names if
Puppet[:immutable_node_data] is set. But the compiler merges in all of
the node.parameters as top scope variables, and these parameters
themselves get the nodes facts merged in too. If the master is using
immutable_node_data (typically has trusted_node_data set to true) then a
node parameter or fact called either 'trusted' or 'facts' will conflict
and raise an error during compiler evaluation as the top scope is
constructed.

Generally this seems like the correct behavior. But PuppetDB stores the
trusted_data as 'trusted' in the node's facts, and if a separate cache
store is not used to preserve the original facts, a Puppet master using
PuppetDB will pull back facts with 'trusted' in them, merge this into a
node's parameters, and fail during compliation.

This patch logs a debug message, but otherwise silently ignores
'trusted' and 'facts' in the node parameters when the compiler merges
them into the top scope.