(PUP-3244) Fixup test for windows and future parser
The test failed on windows nodes because the acceptance test wasn't
using the `puppet` method from beaker. The test failed on future parser
due to the bare file mode, in 4.0 only string modes will be allowed.

This commit fixes up those issues, and removes the
`acceptable_exit_codes => [0]` since that is the default behavior.