(PUP-3277) Use owner from puppet settings rather than beaker defaults
The beaker host['user'] default setting is root, but the actual user
setting from puppet defaults is likely to be puppet or pe-puppet.  By
using the puppet settings, we are more likely to stay in sync with how
the files are being set by packaging or managed by puppet or supporting
puppetlabs modules.  This should cut down on spurious changes to
environment owner/modes that were cropping up in some pe test runs.