(PUP-3277) Do not log certs when maniuplating classifier in acceptance
There is no need to log the master certs when preparing to make
calls to a classifier service on behalf of acceptance tests. Although
the test node certs should not be sensitive, there's still no need to
capture them in the logs, and this commit adds a silent flag to those
config calls.